requirements.txt
pytest>=7.0.0
pillow>=11.1.0
